Add 'state' field to incoming content item
A string. Must be one of 'draft' or 'live'.

Publishing applications can control whether to show content only on the
draft stack by indicating that in the 'state' field:

- draft: will cause content to appear only on the draft site
- live: will cause content to appear on the draft as well as live site.

Publishing API ensures the content ends up in the right content store
which in turn controls which site it'll appear on.