Allow unauthenticated access via content_id header
If a valid JWT token is passed to authenticating-proxy it will extract
the value to the Govuk-Content-Id header, which is passed downstream by
gds-api-adapters.

This change then checks that the header value matches the content ID of
the item being requested, and returns it if so otherwise sends a 403.

Note that the presence of the header means the user has bypassed
normal authentication, so we need to verify that they've specifically
requested the item matching that content_id, and not fall through to
the auth check.

As with a normal access-limited item, we send a cache-control: private
header when the request contains the content-id header.